feat: Add test for custom balance model and interface implementation
