+++ tests/ArchTest.php
@@ -1,5 +1,2 @@
 <?php
 
-it('will not use debugging functions')
-    ->expect(['dd', 'dump', 'ray'])
-    ->each->not->toBeUsed();
